perf(trimul): the channel move back is one kernel, 5.07x on the op and 71.671 -> 67.550 s on the fold
permute(x,(0,2,3,1)) as a single generic_op, replacing the transpose(1,2) + transpose(2,3) pair that
read and wrote the tensor twice. transpose(1,2) on [1,256,512,512] was the largest single class in
the trimul at 512 aa: 4.082 ms at 65.8 GB/s, 17.5 % of the measured combined roof, because the
destination tile spans 32 channels while the source tile spans one. A group (it, jt, ct) takes its 32
input tiles as aligned 2 KB DRAM reads into an L1 window, gathers row il of each as row cl of a
staging tile, the forward direction's compute kernel transposes it, and the writer streams 32 whole
2 KB tiles behind one barrier.

torch.equal at all 30 shapes -- N in {256,288,320,352,384,448,512,576,640,704} x C in {32,64,256} --
against both ttnn.permute and torch.permute (perf/trimul_abs/back_parity_qb2c0.json). At the shape
G=8 runs, [1,256,512,512]: 0.936 ms against the two transposes' 4.745, 5.07x, 287 GB/s, which is
77 % of the measured 375 GB/s combined roof and passes the bytes/time check in both directions.
Speedups run 1.66-5.07x across the sweep.

Fold A/B at 512 aa, qb2 card 0, one process, alternating, warm, both arms at the branch's default
in-projection width so the only thing that moves is the back move
(perf/trimul_abs/fold_e2_512_qb2c0.json): 71.683 / 67.548 / 71.658 / 67.551 s, so
71.671 -> 67.550, 1.0610x, against A/A floors of 0.025 and 0.003 s. TriangleMultiplication over 1208
calls 24.402 -> 20.292 s, 1.2025x, 20.20 -> 16.798 ms per call. The kernel served 1208 of 1208 calls,
0 declined. Bit-exact: CIF sha256 98c33a481fa1fd27 and plDDT 0.828628 on all four arms.

Cumulative with the in-projection gate and the reallocate skip: fold 78.394 -> 67.550 s, 1.1605x, and
the module 31.012 -> 20.292 s per fold, 1.5283x. A re-tape puts the module at 18.617 ms standalone,
2.82x the derived 6.61 ms floor and 1.13x the stock-op floor, with 108 % of the remaining 2.19 ms gap
named in three mechanisms. The forward channel move is 1.62 ms of it and this kernel is the template:
the same gather runs at 287 GB/s in the back direction and 177 in the forward one, where it sits on
the writer interleaved with the DRAM writes and flushed per channel.

Two silent bugs cost three wrong kernels first, and both are general traps for any gather kernel:
get_noc_addr returns uint64_t whose high bits are the target core's x/y, so storing it in a uint32_t
makes noc_async_read_one_packet_set_state name core (0,0) and every gather read fetch another core's
L1 -- wrong at every shape and 5 GB/s, slower than the other bug because the reads become NOC hops.
And noc_async_read_page shares the read command buffer with noc_async_read_one_packet_with_state, so
a kernel doing both must re-issue set_state after every DRAM read batch; setting it once lazily after
the first group's reads is torch.equal TRUE at exactly the shapes where no core owns two groups
(N <= 320 at C=32) and false beyond, which looks like a working kernel.

The eligibility gate is deliberately narrower than the forward one: N must be a multiple of 32 and
the destination must be DRAM. Ragged N keeps the two transposes, since the class this exists for is
the DRAM path from 352 aa up.

Release-gated, stays on this branch.

+++ tt_bio/kernels/reblock_permute_back/reader_reblock_permute_back.cpp
@@ -70,28 +70,30 @@ void kernel_main() {
     const uint32_t end_group = start_group + num_groups;
 
     // The scratch window is this RISC's private staging area, not a producer/consumer queue: the
-    // same kernel writes it and reads it, so it is reserved ONCE and addressed directly. Pushing
-    // and waiting on it per group is what the reference implementation did, and it is wrong here --
-    // a CB's producer and consumer pointers are separate state, so on the second group of a core
-    // `get_read_ptr` no longer names what `get_write_ptr` returned. That failure is invisible while
-    // every core has at most one group (N <= 320 at C = 32, num_groups <= 110) and appears the
-    // moment any core gets two: measured `torch.equal` true at N = 256/288/320 with C = 32 and
-    // false at every larger cell, with the rate collapsing to 8-13 GB/s on the semaphore round
-    // trips as well. The forward direction's writer already uses this pattern for its staging CB.
+    // same kernel writes it and reads it, so it is reserved ONCE and addressed directly, the way the
+    // forward direction's writer holds its own staging CB.
     cb_reserve_back(scratch_cb_id, TILE_HEIGHT);
     const uint32_t group_l1_base = get_write_ptr(scratch_cb_id);
 
-    // Every gather read is local L1 -> L1 at exactly FACE_ROW_BYTES, so the NOC coordinates and the
-    // length are the same for all of them and can be written into the read command buffer once per
-    // group instead of once per transaction. ONCE PER GROUP AND NOT ONCE PER KERNEL: this kernel
-    // also issues `noc_async_read_page` for the group's DRAM tiles, and that shares the same read
-    // command buffer, so the DRAM reads overwrite the one-packet state. Setting it at kernel entry
-    // makes every gather read use a DRAM page's NOC coordinates -- measured as `torch.equal` false
-    // at every shape and a uniform 8-9 GB/s. Setting it once and then reading DRAM per group is
-    // worse than useless, it is correct only for the first group of each core, which passes at
-    // N <= 320, C = 32 and fails everywhere a core owns two groups. The forward direction's writer
-    // gets to set it once because a writer issues no reads other than its own gather.
-    const uint32_t gather_state_addr = get_noc_addr(group_l1_base);
+    // Every gather read is local L1 -> L1 at exactly FACE_ROW_BYTES, so the coordinates and the
+    // length are the same for all of them and go into the read command buffer once per GROUP rather
+    // than once per transaction.
+    //
+    // Two things about this line were each worth a wrong kernel, and both are silent:
+    //
+    //   * it is uint64_t, not uint32_t. `get_noc_addr` returns a 64-bit NOC address whose high bits
+    //     are the target core's x/y; truncating it to 32 bits keeps only the local offset, so the
+    //     state names core (0,0) and every gather read fetches another core's L1. Measured
+    //     `torch.equal` false at every shape at a uniform 5 GB/s -- SLOWER than the wrong-state case
+    //     below, because the reads become real NOC hops.
+    //   * it is re-issued per group and not once per kernel. This kernel also calls
+    //     `noc_async_read_page` for the group's DRAM tiles and that shares the same read command
+    //     buffer, so the DRAM reads overwrite the one-packet state. Setting it once at entry
+    //     measured false everywhere at 9 GB/s; setting it once lazily after the first group's reads
+    //     measured TRUE at exactly the shapes where no core owns two groups (N <= 320 at C = 32) and
+    //     false beyond, which is the shape-dependent failure that makes this trap dangerous. The
+    //     forward writer gets to set it once because a writer issues no reads but its own gather.
+    const uint64_t gather_state_addr = get_noc_addr(group_l1_base);
 
     for (uint32_t group = start_group; group < end_group; ++group) {
         const uint32_t it = group / NtCt;
